% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.vector_cache_path}
\alias{.vector_cache_path}
\title{Return the default on-disk path for a cached model file}
\usage{
.vector_cache_path(
  model = "glove.6B.50d",
  cache_dir = rappdirs::user_data_dir("lexiExpand")
)
}
\arguments{
\item{model}{Character. Model name, e.g. \code{"glove.6B.50d"}.}

\item{cache_dir}{Character. Root cache directory.}
}
\value{
Character scalar: the full path to the \code{.txt} vector file.
}
\description{
Return the default on-disk path for a cached model file
}
\keyword{internal}
