% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_candidates.R
\name{print.lexiexpand_result}
\alias{print.lexiexpand_result}
\title{Print an lexiexpand_result}
\usage{
\method{print}{lexiexpand_result}(x, ...)
}
\arguments{
\item{x}{A \code{lexiexpand_result} object.}

\item{...}{Ignored.}
}
\value{
\code{x} invisibly.
}
\description{
Print an lexiexpand_result
}
